requirements.txt
python-telegram-bot[job-queue]==21.6
python-dotenv==1.0.1
aiosqlite==0.20.0
asyncpg==0.29.0
openpyxl==3.1.2
